+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file
 ### Fixed
 
 - App crashed when dealing with expired PGP keys. From now on, password items can be decrypted and displayed even after the key has expired, but they can no longer be edited, and no new items can be created.
+- Handling of password items that were encrypted to multiple subkeys
 - Extra content: strip leading whitespace when copied to clipboard but preserve indentation of multiline text; append newline character at last line
 
 ## [1.15.4] - 2025-11-27